[feat] : 포인트 관리자 및 마이페이지 조회 api 작성완료

+++ src/main/java/shop/itbook/itbookshop/pointgroup/pointincreasedecreasecontent/repository/PointIncreaseDecreaseContentRepository.java
@@ -9,6 +9,6 @@ import shop.itbook.itbookshop.pointgroup.pointincreasedecreasecontent.entity.Poi
  * @since 1.0
  */
 public interface PointIncreaseDecreaseContentRepository
-    extends JpaRepository<PointIncreaseDecreaseContent, Long> {
+    extends JpaRepository<PointIncreaseDecreaseContent, Integer> {
 
 }
